+++ src/CodeWF.AvaloniaControls.Showcase/Views/MainWindow.axaml.cs
@@ -1,13 +1,11 @@
-using CodeWF.AvaloniaControls.Helpers;
-using Ursa.Controls;
+using CodeWF.AvaloniaControls.Controls;
 
 namespace CodeWF.AvaloniaControls.Showcase.Views;
 
-public partial class MainWindow : UrsaWindow
+public partial class MainWindow : CodeWFWindow
 {
     public MainWindow()
     {
         InitializeComponent();
-        this.EnableOSVersionAwareDecorations();
     }
 }