+++ src/CodeWF.AvaloniaControls.Showcase/Views/TestNoneWindowDemos/CodeWFWindowAccentDemo.axaml.cs
@@ -1,0 +1,11 @@
+using CodeWF.AvaloniaControls.Controls;
+
+namespace CodeWF.AvaloniaControls.Showcase.Views.TestNoneWindowDemos;
+
+public partial class CodeWFWindowAccentDemo : CodeWFWindow
+{
+    public CodeWFWindowAccentDemo()
+    {
+        InitializeComponent();
+    }
+}