+++ src/CodeWF.AvaloniaControls.Showcase/Views/TestNoneWindowDemos/CodeWFWindowCompactToolDemo.axaml.cs
@@ -1,0 +1,11 @@
+using CodeWF.AvaloniaControls.Controls;
+
+namespace CodeWF.AvaloniaControls.Showcase.Views.TestNoneWindowDemos;
+
+public partial class CodeWFWindowCompactToolDemo : CodeWFWindow
+{
+    public CodeWFWindowCompactToolDemo()
+    {
+        InitializeComponent();
+    }
+}