+++ src/CodeWF.AvaloniaControls.Showcase/Views/TestNoneWindowDemos/CodeWFWindowTallDemo.axaml.cs
@@ -1,0 +1,11 @@
+using CodeWF.AvaloniaControls.Controls;
+
+namespace CodeWF.AvaloniaControls.Showcase.Views.TestNoneWindowDemos;
+
+public partial class CodeWFWindowTallDemo : CodeWFWindow
+{
+    public CodeWFWindowTallDemo()
+    {
+        InitializeComponent();
+    }
+}