+++ src/CodeWF.AvaloniaControls.Showcase/Views/TestNoneWindowDemos/UrsaWindowCompactToolDemo.axaml.cs
@@ -1,0 +1,11 @@
+using Ursa.Controls;
+
+namespace CodeWF.AvaloniaControls.Showcase.Views.TestNoneWindowDemos;
+
+public partial class UrsaWindowCompactToolDemo : UrsaWindow
+{
+    public UrsaWindowCompactToolDemo()
+    {
+        InitializeComponent();
+    }
+}