+++ src/CodeWF.AvaloniaControls.Showcase/Views/TestNoneWindowDemos/UrsaWindowDarkTitleBarDemo.axaml.cs
@@ -1,0 +1,11 @@
+using Ursa.Controls;
+
+namespace CodeWF.AvaloniaControls.Showcase.Views.TestNoneWindowDemos;
+
+public partial class UrsaWindowDarkTitleBarDemo : UrsaWindow
+{
+    public UrsaWindowDarkTitleBarDemo()
+    {
+        InitializeComponent();
+    }
+}