+++ src/CodeWF.AvaloniaControls.Showcase/Views/TestNoneWindowDemos/UrsaWindowShapedDemo.axaml.cs
@@ -1,0 +1,26 @@
+using Avalonia.Input;
+using Avalonia.Interactivity;
+using Ursa.Controls;
+
+namespace CodeWF.AvaloniaControls.Showcase.Views.TestNoneWindowDemos;
+
+public partial class UrsaWindowShapedDemo : UrsaWindow
+{
+    public UrsaWindowShapedDemo()
+    {
+        InitializeComponent();
+    }
+
+    private void TitleBar_OnPointerPressed(object? sender, PointerPressedEventArgs e)
+    {
+        if (e.GetCurrentPoint(this).Properties.IsLeftButtonPressed)
+        {
+            BeginMoveDrag(e);
+        }
+    }
+
+    private void Close_OnClick(object? sender, RoutedEventArgs e)
+    {
+        Close();
+    }
+}