+++ src/CodeWF.AvaloniaControls.Showcase/Views/TestNoneWindowDemos/UrsaWindowTallTitleBarDemo.axaml.cs
@@ -1,0 +1,11 @@
+using Ursa.Controls;
+
+namespace CodeWF.AvaloniaControls.Showcase.Views.TestNoneWindowDemos;
+
+public partial class UrsaWindowTallTitleBarDemo : UrsaWindow
+{
+    public UrsaWindowTallTitleBarDemo()
+    {
+        InitializeComponent();
+    }
+}