+++ src/CodeWF.AvaloniaControls.Showcase/Views/TestNoneWindowDemos/UrsaWindowWithNoneAndMove.axaml.cs
@@ -1,4 +1,5 @@
-using CodeWF.AvaloniaControls.Helpers;
+using Avalonia.Input;
+using Avalonia.Interactivity;
 using Ursa.Controls;
 
 namespace CodeWF.AvaloniaControls.Showcase.Views.TestNoneWindowDemos;
@@ -8,6 +9,18 @@ public partial class UrsaWindowWithNoneAndMove : UrsaWindow
     public UrsaWindowWithNoneAndMove()
     {
         InitializeComponent();
-        this.EnableOSVersionAwareDecorations();
+    }
+
+    private void TitleBar_OnPointerPressed(object? sender, PointerPressedEventArgs e)
+    {
+        if (e.GetCurrentPoint(this).Properties.IsLeftButtonPressed)
+        {
+            BeginMoveDrag(e);
+        }
+    }
+
+    private void Close_OnClick(object? sender, RoutedEventArgs e)
+    {
+        Close();
     }
 }